feature Invoice: split EditInvoice in multiple components: EditInvoiceHeader, EditInvoiceBody, EditInvoiceFooter
