+++ frontend/src/components/invoice/invoice-edit/EditInvoice.tsx
@@ -1,35 +1,21 @@
 import {useEffect, useReducer, useState} from 'react';
-import {useDispatch, useSelector} from 'react-redux';
-import {Container, Row, Col, Form} from 'react-bootstrap';
+import {useSelector} from 'react-redux';
+import {Container, Row, Form} from 'react-bootstrap';
 import {t} from '../../utils';
-import {EditInvoiceLines} from './invoice-lines/EditInvoiceLines';
-import InvoiceNotVerifiedAlert from './InvoiceNotVerifiedAlert';
-import {EditInvoiceSaveButtons} from './EditInvoiceSaveButtons';
-import {createInvoice, previewInvoice, syncCreditNotas, updateInvoiceRequest} from '../../../actions/index';
-import {EditInvoiceClient} from './EditInvoiceClient';
 import InvoiceModel from '../models/InvoiceModel';
 import {ConfacState} from '../../../reducers/app-state';
-import {EditInvoiceDetails} from './EditInvoiceDetails';
 import {StickyFooter} from '../../controls/other/StickyFooter';
-import {DownloadInvoiceButton} from './DownloadInvoiceButton';
 import {EmailModal, EmailTemplate} from '../../controls/email/EmailModal';
-import {Button} from '../../controls/form-controls/Button';
-import {getNewClonedInvoice, getNewInvoice} from '../models/getNewInvoice';
 import {useDocumentTitle} from '../../hooks/useDocumentTitle';
-import {InvoiceAttachmentsForm} from '../controls/InvoiceAttachmentsForm';
-import {EditInvoiceBadges} from './EditInvoiceBadges';
-import {Audit} from '../../admin/audit/Audit';
-import {Claim} from '../../users/models/UserModel';
-import {useProjectsMonth} from '../../hooks/useProjects';
 import {useParams} from 'react-router-dom';
-import {ProjectMonthOrManualSelect} from '../../project/controls/ProjectMonthOrManualSelect';
-import {InvoiceDownloadIcon} from '../../controls/attachments/AttachmentDownloadIcon';
 import useEntityChangedToast from '../../hooks/useEntityChangedToast';
 
 
 import './EditInvoice.scss';
-import { InvoiceCreditNotas } from '../controls/InvoiceCreditNotas';
-import { NotesWithCommentsModalButton } from '../../controls/form-controls/button/NotesWithCommentsModalButton';
+import { EditInvoiceHeader } from './EditInvoiceHeader';
+import { EditInvoiceBody } from './EditInvoiceBody';
+import { EditInvoiceFooter } from './EditInvoiceFooter';
+import { getNewInvoice } from '../models/getNewInvoice';
 
 
 const EditInvoice = () => {
@@ -45,11 +31,10 @@ const EditInvoice = () => {
   const invoices = useSelector((state: ConfacState) => state.invoices);
   const clients = useSelector((state: ConfacState) => state.clients);
   const initInvoice = storeInvoice ? new InvoiceModel(config, storeInvoice) : getNewInvoice(config, invoices, clients, {isQuotation});
-  const fullProjectMonth = useProjectsMonth(storeInvoice?.projectMonth?.projectMonthId);
+
   const [invoice, setInvoice] = useState<InvoiceModel>(initInvoice);
   useEntityChangedToast(invoice._id);
   const [, forceUpdate] = useReducer(x => x + 1, 0);
-  const dispatch = useDispatch();
 
   // useEffect(() => window.scrollTo(0, 0)); // TODO: each keystroke made it scroll to top :(
   const [showEmailModal, setEmailModal] = useState<EmailTemplate>(EmailTemplate.None);
@@ -72,16 +57,6 @@ const EditInvoice = () => {
     setInvoice(new InvoiceModel(config, navigateInvoice))
   }, [params, config, invoices])
 
-  const type: 'quotation' | 'invoice' = isQuotation ? 'quotation' : 'invoice';
-
-  const updateInvoice = (key: string, value: any, calcMoneys = false) => {
-    // Naughty naughty: We are manipulating state directly!
-    // To fix this: state should be a regular object, and a
-    // InvoiceModel should be created in the render
-    invoice.updateField(key, value, calcMoneys);
-    setInvoice(invoice);
-    forceUpdate();
-  }
 
   // TODO: confusion with storeInvoice vs initInvoice vs invoice
   // --> There should be a form variant and a model variant new'd for every render
@@ -93,73 +68,23 @@ const EditInvoice = () => {
     <Container className="edit-container">
       <Form>
         <Row>
-          <Col sm={12} style={{marginBottom: 20, display: 'flex', alignItems: 'center', justifyContent: 'space-between'}}>
-            <div style={{display: 'inline-flex', alignItems: 'flex-start'}}>
-              <h1 style={{width: 'unset'}}>
-                {initInvoice._id ? t(`${type}.editTitle`) : t(`${type}.createTitle`)}
-                <Audit model={storeInvoice} modelType="invoice" />
-              </h1>
-              <div>
-                <EditInvoiceBadges invoice={invoice} />
-              </div>
-            </div>
-            <div>
-              <div className={`invoice-top-buttonbar ${storeInvoice?._id ? 'invoice-edit' : 'invoice-new'}`}>
-                {<NotesWithCommentsModalButton
-                  claim={invoice.isQuotation ? Claim.ManageQuotations : Claim.ManageInvoices}
-                  value={{note: invoice.note, comments: invoice.comments || []}}
-                  onChange={val => {
-                    setInvoice(new InvoiceModel(invoice.config, {...invoice, note: val.note, comments: val.comments}));
-                    forceUpdate();
-                  }}
-                  title={t('projectMonth.note')}
-                  variant="link"
-                />}
-                {initInvoice._id && <DownloadInvoiceButton invoice={initInvoice} />}
-                {storeInvoice?._id && !storeInvoice?.isQuotation && <InvoiceDownloadIcon invoice={invoice} fileType='xml' style={{color: '#0062cc', marginLeft: 20}} />}
-              </div>
-            </div>
-          </Col>
-          <Col sm={12}>
-            <InvoiceNotVerifiedAlert invoice={initInvoice} />
-          </Col>
+          <EditInvoiceHeader
+            invoice={invoice}
+            isNew={!initInvoice._id}
+            onChange={invoice => {
+              setInvoice(invoice)
+              forceUpdate()
+            }}
+          />
         </Row>
-
         <Row>
-          <Col sm={6}>
-            <EditInvoiceClient
-              invoice={invoice}
-              onChange={val => {
-                setInvoice(invoice.setClient(val))
-                forceUpdate();
-              }}
-            />
-          </Col>
-
-          <Col sm={6}>
-            <Row>
-              <EditInvoiceDetails
-                invoice={invoice}
-                onChange={(fieldName: string, value: any) => updateInvoice(fieldName, value, true)}
-              />
-            </Row>
-            <Row>
-              <ProjectMonthOrManualSelect
-                value={invoice.projectMonth}
-                onProjectMonthChange={fpm => {
-                  invoice.setProjectMonth(fpm);
-                  setInvoice(invoice);
-                  forceUpdate();
-                }}
-                onManualChange={(consultant, month) => {
-                  invoice.setManualProjectMonth(consultant, month || undefined);
-                  setInvoice(invoice);
-                  forceUpdate();
-                }}
-                invoice={invoice}
-              />
-            </Row>
-          </Col>
+          <EditInvoiceBody
+            invoice={invoice}
+            onChange={invoice => {
+              setInvoice(invoice)
+              forceUpdate()
+            }}
+          />
         </Row>
 
         {!!initInvoice._id && invoice.client && showEmailModal !== EmailTemplate.None && (
@@ -170,70 +95,12 @@ const EditInvoice = () => {
           />
         )}
 
-        <Row style={{marginTop: 8}}>
-          <EditInvoiceLines
-            claim={invoice.isQuotation ? Claim.ManageQuotations : Claim.ManageInvoices}
-            value={invoice.lines}
-            onChange={m => {
-              setInvoice(invoice.setLines(m));
-              forceUpdate();
-            }}
-            translationPrefix={invoice.getType()}
-          />
-        </Row>
-        <InvoiceCreditNotas
-          model={invoice}
-          onChange={m => {
-            setInvoice(m)
-            forceUpdate()
-          }}
-        />
-        <InvoiceAttachmentsForm model={initInvoice} />
+
         <StickyFooter>
-          {!initInvoice.isNew && (
-            <>
-              <Button
-                claim={invoice.isQuotation ? Claim.ManageQuotations : Claim.EmailInvoices}
-                variant={storeInvoice?.verified || storeInvoice?.lastEmail ? 'outline-danger' : 'light'}
-                icon="far fa-envelope"
-                onClick={() => setEmailModal(EmailTemplate.InitialEmail)}
-                className="tst-open-email-initial"
-              >
-                {t('email.prepareEmail')}
-              </Button>
-              <Button
-                claim={invoice.isQuotation ? Claim.ManageQuotations : Claim.EmailInvoices}
-                variant={storeInvoice?.verified || !storeInvoice?.lastEmail ? 'outline-danger' : 'light'}
-                icon="far fa-envelope"
-                onClick={() => setEmailModal(EmailTemplate.Reminder)}
-                className="tst-open-email-reminder"
-              >
-                {t('email.prepareEmailReminder')}
-              </Button>
-            </>
-          )}
-          <EditInvoiceSaveButtons
-            onClick={(type, navigate) => {
-              // When the InvoiceNotVerifiedAlert component updates the verified flag to true on initInvoice,
-              // it doesn't update on the edited object invoice. So the verified flag is overridden when you save
-              // the edited object invoice. This if check fixes that
-              if (initInvoice.verified !== invoice.verified) {
-                invoice.verified = initInvoice.verified;
-              }
-              if (type === 'create') {
-                dispatch(createInvoice(invoice, navigate) as any);
-              } if (type === 'preview') {
-                dispatch(previewInvoice(invoice.client.invoiceFileName || config.invoiceFileName, invoice, fullProjectMonth) as any);
-              } if (type === 'update') {
-                dispatch(syncCreditNotas(invoice, initInvoice.creditNotas, invoices) as any)
-                dispatch(updateInvoiceRequest(invoice, undefined, false, navigate) as any);
-              } if (type === 'clone') {
-                const creditNota = getNewClonedInvoice(invoices, invoice)
-                dispatch(syncCreditNotas(creditNota, initInvoice.creditNotas, invoices) as any)
-                dispatch(createInvoice(creditNota, navigate) as any);
-              }
-            }}
-            invoice={initInvoice}
+          <EditInvoiceFooter
+            invoice={invoice}
+            initInvoice={initInvoice}
+            setEmailModal={setEmailModal}
           />
         </StickyFooter>
       </Form>
